Minor bug and damage states

+++ Assets/Scripts/UI/Oxygen.cs
@@ -29,12 +29,13 @@ public class Oxygen : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
+        
         if (gameObject.GetComponent<Image>().IsActive())
         {
             timeLeft -= Time.deltaTime;
             SetValue(timeLeft / timeLimit);
         }
-        if (timeLeft <= 0)
+        if (timeLeft <= 0 && Health.GetInstance().getCurrHealth() > 0)
         {
             // implement fail state
             Health.GetInstance().kill();
